# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API文档",
        "title": "天气管理系统后端API文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:9027",
    "basePath": "/",
    "paths": {
        "/data": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据管理"
                ],
                "summary": "显示当前用户可见的所有数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前页数,默认为1",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页大小，默认为7",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前用户id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "条件筛选: 设备ID",
                        "name": "nodeID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "条件筛选: 物理量",
                        "name": "dataName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "条件筛选: 时间:yy.mm.dd hh:mm:ss - yy.mm.dd hh:mm:ss",
                        "name": "time",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/data/download/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据管理"
                ],
                "summary": "导出当前用户的所有可见设备的数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "当前用户id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "条件筛选: 设备ID",
                        "name": "nodeID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "条件筛选: 物理量",
                        "name": "dataName",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "条件筛选: 时间:yy.mm.dd hh:mm:ss - yy.mm.dd hh:mm:ss",
                        "name": "time",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "下载成功",
                        "schema": {
                            "$ref": "#/definitions/controller.SwaggerResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.SwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer"
                },
                "msg": {
                    "description": "描述",
                    "type": "string"
                }
            }
        },
        "controller.SwaggerResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "msg": {
                    "description": "描述",
                    "type": "string"
                }
            }
        },
...
    }
}